+++ lib/interface.ts
@@ -100,6 +100,11 @@ export interface IAjaxProcessDataOptions {
     reject?: IReject;
 }
 
+export interface IAjaxProcessDataAfterOptions extends IAjaxProcessDataOptions {
+    /** 为false时不格式化请求参数 */
+    processData?: boolean;
+}
+
 export interface IProcessResponseOptions extends IAjaxArgsOptions {
     xhr: XMLHttpRequest;
     reject: IReject;
@@ -261,6 +266,10 @@ export interface IAjax {
     processData: (
         params: IParams,
         props: { method: IMethod; url: string; options: IOptions; reject?: IReject }
+    ) => IParams;
+    processDataAfter: (
+        params: IParams,
+        props: { method: IMethod; url: string; options: IOptions; reject?: IReject; processData?: boolean }
     ) => IParams;
     processResponse: (response: IResult | null, props: IProcessResponseOptions) => IResult;
     readonly stringifyParams: (